feat: improved packet events
