+++ src/main/java/me/adrianed/unsignedvelocity/event/PacketEvent.java
@@ -1,0 +1,22 @@
+package me.adrianed.unsignedvelocity.event;
+
+import com.velocitypowered.api.proxy.Player;
+import com.velocitypowered.proxy.protocol.MinecraftPacket;
+
+public abstract class PacketEvent {
+    private final MinecraftPacket packet;
+    private final Player player;
+
+    protected PacketEvent(MinecraftPacket packet, Player player) {
+        this.packet = packet;
+        this.player = player;
+    }
+
+    public MinecraftPacket getPacket() {
+        return this.packet;
+    }
+
+    public Player getPlayer() {
+        return this.player;
+    }
+}
